- made compilable with FreePacsal 2.0 - removed some warnings about variable resizing with possible loss of information
git-svn-id: https://svn.openxp.de/openxp/trunk@6899 65356fb5-50e6-0310-a06b-fbc0adf8d39c

+++ fidoglob.pas
@@ -294,7 +294,8 @@ begin
 end;
 
 function  TNodeList.GetFileName(n:integer):string;
-var p : byte;
+var
+  p: Integer;
 begin
   if n>=Count then
     result:=''
