+++ lister.pas
@@ -551,7 +551,7 @@ var
   procedure suchen(rep: boolean);
   var
     found, brk: boolean;
-    pp: byte;
+    pp: Integer;
     sw: byte;
     nftxt: atext;
     mi: byte;
